Team Funding Request 3 (#18)
* Team Funding Request #3

# Summary

This proposal is a funding request by the Particl team. It requests **35,250 PART per month** from the treasury fund for a period of **12 months** after which the team will submit a new funding request for a community vote.

### Note

This proposal is a brief, summarized version of our extensive proposal available [here](https://particl.news/particl-team-funding-request-3/). If you haven't yet done it, please take the time to read the full proposal, which contains all of the details surrounding this funding request, as well as our [newly-announced strategic path forward](https://particl.news/particls-next-evolutionary-leap-forward/) which we'll be following all throughout this funding cycle.

# Use of Funds

The Particl team operates strictly on community donations under the approval of Particl stakeholders. Funding, in PART, is taken directly from the network’s treasury and allocated to the team to fund its operations. 

With the current PART price in mind, **the Particl team is requesting 35,250 PART coins per month, for 12 months**, from the treasury to uninterruptedly fund the upcoming development cycle at the current pace and pay for the various operation-related expenses, incuding legal, administration, corporate fees, and similar spendings. 

With better PART prices and liquidity, the extra PART will be used to grow the team and fund various marketing, communications, community, and outreach efforts to promote the ecosystem and its dApps.

# Legal Entity

All funds voted to be allocated to the Particl team are held and managed by SoonTM Development Ltd ([SoonTM.io](https://soontm.io), an independent software development company.

# How to Vote

* To learn more about Particl's on-chain voting system, [click here](https://academy.particl.io/en/latest/in-depth/indepth_voting.html#).
* To learn how to vote, refer to [this Particl Wiki tutorial](https://particl.wiki/tutorial/staking/how-to-vote/).

To ensure the proposal is properly showing up on your Partyman app, you need to update it to the latest version. To do so, run `git pull` from inside your partyman folder, and then type the `./partyman update` command from that same partyman folder.

## Read More

* To learn more details about the team's proposal, please read the full proposal available [here](https://particl.news/particl-team-funding-request-3/).
* To know more about the team's strategy moving forward, read the team's major development announcement available [here](https://particl.news/particls-next-evolutionary-leap-forward/).

* Team Funding Request #3

# Summary

This proposal is a funding request by the Particl team. It requests **35,250 PART per month** from the treasury fund for a period of **12 months** after which the team will submit a new funding request for a community vote.

### Note

This proposal is a brief, summarized version of our extensive proposal available [here](https://particl.news/particl-team-funding-request-3/). If you haven't yet done it, please take the time to read the full proposal, which contains all of the details surrounding this funding request, as well as our [newly-announced strategic path forward](https://particl.news/particls-next-evolutionary-leap-forward/) which we'll be following all throughout this funding cycle.

# Use of Funds

The Particl team operates strictly on community donations under the approval of Particl stakeholders. Funding, in PART, is taken directly from the network’s treasury and allocated to the team to fund its operations. 

With the current PART price in mind, **the Particl team is requesting 35,250 PART coins per month, for 12 months**, from the treasury to uninterruptedly fund the upcoming development cycle at the current pace and pay for the various operation-related expenses, incuding legal, administration, corporate fees, and similar spendings. 

With better PART prices and liquidity, the extra PART will be used to grow the team and fund various marketing, communications, community, and outreach efforts to promote the ecosystem and its dApps.

# Legal Entity

All funds voted to be allocated to the Particl team are held and managed by SoonTM Development Ltd ([SoonTM.io](https://soontm.io), an independent software development company.

# How to Vote

* To learn more about Particl's on-chain voting system, [click here](https://academy.particl.io/en/latest/in-depth/indepth_voting.html#).
* To learn how to vote, refer to [this Particl Wiki tutorial](https://particl.wiki/tutorial/staking/how-to-vote/).

To ensure the proposal is properly showing up on your Partyman app, you need to update it to the latest version. To do so, run `git pull` from inside your partyman folder, and then type the `./partyman update` command from that same partyman folder.

## Read More

* To learn more details about the team's proposal, please read the full proposal available [here](https://particl.news/particl-team-funding-request-3/).
* To know more about the team's strategy moving forward, read the team's major development announcement available [here](https://particl.news/particls-next-evolutionary-leap-forward/).

* Rename team-funding-request-3 to team-funding-request-3.md

* Update team-funding-request-3.md

* Update team-funding-request-3.md

* Update team-funding-request-3.md

* Minor formatting edit